show stats panel more reliably
with the new contact info panel, the stats panel was once again
showing up with ugly results. it turns out that the order of things
really really matters: the stats panel going last with the store
after the view makes it 100% reliable.

while in there i also improved some of the text a bit and shortened
the contact pane's contact by making the warehouse name a link